- Fix date syncing. - Create test for other date bug. - Update script, including logging - Update version. - Update README.md - Add log statement. - Fix some warnings. - Remove unneeded getters.
