+++ .codacy.yml
@@ -1,0 +1,3 @@
+---
+exclude_paths:
+  - "**/test/**"
