+++ projectHypeBerry.egg-info/PKG-INFO
@@ -1,6 +1,6 @@
 Metadata-Version: 1.0
 Name: Project_Hype-Berry
-Version: 2.1.0
+Version: 2.1.1
 Summary: An API app for syncing todoist and habitica tasks
 Home-page: https://github.com/programmerPhysicist/Project_Hype-Berry
 Author: UNKNOWN
