+++ scripts/debugTests.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+
+pytest --pdb --pdbcls=debugpy:launcher
