+++ setup.py
@@ -4,7 +4,7 @@ from setuptools import setup
 
 setup(
     name='Project_Hype-Berry',
-    version='2.1.0',
+    version='2.1.1',
     url='https://github.com/programmerPhysicist/Project_Hype-Berry',
     description='An API app for syncing todoist and habitica tasks',
     packages=['Project_Hype-Berry'],
